VoiceViz_backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
SQLAlchemy==2.0.40
PyMySQL==1.1.1
passlib==1.7.4
bcrypt==4.3.0
python-jose==3.4.0
dnspython==2.7.0
python-multipart==0.0.20
